pyproject.toml

[project]
name = "Fyyur"
version = "0.0"
authors = []
dependencies = [
    "babel",
    "flask",
    "flask_migrate",
    "flask-moment",
    "flask_sqlalchemy",
    "flask-wtf",
    "psycopg2",
    "pydantic",
    "python-dateutil",
]

[project.optional-dependencies]
dev = ["black", "pytest", "ruff"]

[tool.ruff]
select = ["E", "F", "I"]
line-length = 90
exclude = ["fabfile.py"]
fixable = ["ALL"]

[tool.black]
line-length = 90

[tool.setuptools]
py-modules = []
